+++ app/services/create_journey_step.rb
@@ -41,6 +41,7 @@ class CreateJourneyStep
       contentful_type: step_type,
       options: options,
       primary_call_to_action_text: primary_call_to_action_text,
+      hidden: hidden,
       raw: raw,
       journey: journey
     )
@@ -98,6 +99,12 @@ class CreateJourneyStep
   def primary_call_to_action_text
     return nil unless contentful_entry.respond_to?(:primary_call_to_action)
     contentful_entry.primary_call_to_action
+  end
+
+  def hidden
+    return false unless contentful_entry.respond_to?(:always_show_the_user)
+    return false if contentful_entry.always_show_the_user.nil?
+    !contentful_entry.always_show_the_user
   end
 
   def raw